Wendland functions using Chebyshev polynomials.
The evaluation routines of the Wendland functions have been changed. Instead
of the monomial basis, the evaluation is carried out in the much more stable
Chebyshev basis. The coefficients are computed in exact, rational arithmetic
and only converted to double as the last final step. Evaluation is done using
the Clenshaw algorithm.